reorganisation des vues et corrections templates
