+++ templates/tutorial/chapter/view.html
@@ -34,7 +34,7 @@
 
         {% include 'tutorial/includes/tags_authors.part.html' %}
 
-        {% if tutorial.in_beta and tutorial.sha_beta == version %}
+        {% if tutorial.in_beta %}
             <div class="content-wrapper">
                 <div class="alert-box warning">
                     Attention, cette version du tutoriel est en BETA !
