+++ zds/pages/views.py
@@ -28,8 +28,8 @@ def home(request):
 
     tutos = []
     for tuto in get_last_tutorials():
-        data = tuto.load_json_for_public()
-        data = tuto.load_dic(data)
+        data = tuto.load_json(tuto.sha_public)
+        data = tuto.load_data(mandata=data, public =True)
         tutos.append(data)
     
     articles = []
