+++ zds/tutorial/views/__init__.py
@@ -1,0 +1,8 @@
+from views_index import *
+from views_tutorial import *
+from views_part import *
+from views_chapter import *
+from views_extract import *
+from views_note import *
+from validation import *
+from versionning import *
